Implement self-initializing TObject wrappers
gPad is NULL at Perl compile time. Thus we have to jump through a lot of
hoops to provide a Perl-global $gPad. Instead of instantiating a normal
TObject wrapper via EncapsulateObject, we create a wrapper that contains
a pointer to the TObject* pointer instead. This wrapper is made
"magical" (special Perl guts term, see "perldoc perlguts") and
automatically dereferences itself on first use. This cost me a bit of my
sanity to implement.