First (broken) stab at implementing reassignment of global ROOT variables
Sadly, this doesn't work.

If we do gRandom = new TRandom(123) at the C level, right now, this
means that we're now tracking the wrong C variable. Some ROOT functions
will be refering to "gRandom" so we'd have to track &gRandom instead of
the content of gRandom itself (since that pointer's now changed). Needs
more thinking. A lot more.